+++ src/routes/datasets/Datasets.tsx
@@ -93,20 +93,13 @@ const Datasets = (): JSX.Element => {
                                     width="1px"
                                     whiteSpace="nowrap"
                                 >
-                                    Processable by
-                                </Th>
-                                <Th
-                                    textAlign="right"
-                                    width="1px"
-                                    whiteSpace="nowrap"
-                                >
-                                    Downloadable by
+                                    Permissions
                                 </Th>
                             </Tr>
                         </Thead>
                         <Tbody data-cy={datasetsLoading ? 'loading' : 'loaded'}>
                             {!datasetsLoading && datasets.length === 0 && (
-                                <EmptyTr nbColumns={3} />
+                                <EmptyTr nbColumns={2} />
                             )}
                             {datasetsLoading ? (
                                 <TableSkeleton
@@ -124,9 +117,6 @@ const Datasets = (): JSX.Element => {
                                                 by Foo
                                             </Text>
                                         </Skeleton>
-                                    </Td>
-                                    <Td textAlign="right">
-                                        <Skeleton width="100px" height="20px" />
                                     </Td>
                                     <Td textAlign="right">
                                         <Skeleton width="100px" height="20px" />
@@ -156,13 +146,6 @@ const Datasets = (): JSX.Element => {
                                             <PermissionTag
                                                 permission={
                                                     dataset.permissions.process
-                                                }
-                                            />
-                                        </Td>
-                                        <Td textAlign="right">
-                                            <PermissionTag
-                                                permission={
-                                                    dataset.permissions.download
                                                 }
                                             />
                                         </Td>